License fix: Move messageconverter to common
