+++ src/libertem/common/messageconverter.py
