+++ src/libertem/io/dataset/base/dataset.py
@@ -7,7 +7,7 @@ from libertem.common.shape import Shape
 from libertem.common.math import prod
 from libertem.io.utils import get_partition_shape
 from libertem.io.dataset.base import DataSetException, MMapBackend
-from libertem.web.messageconverter import MessageConverter
+from libertem.common.messageconverter import MessageConverter
 from libertem.io.corrections.corrset import CorrectionSet
 from .partition import BasePartition, Partition
