+++ src/libertem/io/dataset/frms6.py
@@ -15,7 +15,7 @@ import numpy as np
 from libertem.common.math import prod
 from libertem.io.corrections import CorrectionSet
 from libertem.common import Shape, Slice
-from libertem.web.messages import MessageConverter
+from libertem.common.messageconverter import MessageConverter
 from .base import (
     DataSet, DataSetException, DataSetMeta,
     FileSet, BasePartition, File, Decoder,
