+++ src/libertem/io/dataset/raw.py
@@ -4,7 +4,7 @@ import numpy as np
 
 from libertem.common.math import prod
 from libertem.common import Shape
-from libertem.web.messages import MessageConverter
+from libertem.common.messageconverter import MessageConverter
 from .base import (
     DataSet, DataSetException, DataSetMeta,
     BasePartition, File, FileSet, DirectBackend, IOBackend,
