+++ src/libertem/io/dataset/tvips.py
@@ -9,7 +9,7 @@ from glob import glob, escape
 from libertem.common.math import prod
 from libertem.common import Shape
 from libertem.executor.base import JobExecutor
-from libertem.web.messages import MessageConverter
+from libertem.common.messageconverter import MessageConverter
 from .base import (
     DataSet, DataSetException, DataSetMeta,
     BasePartition, File, FileSet, IOBackend,
